Move query/remove/importer into own directories; update main script; remove obsolete READMEs; update CHANGELOG
