--- src/lib/README.md
@@ -1,7 +1,0 @@
-# lib
-This folder should contain nothing but this and other folders.
-- In here is a general organization of the scripts that do the heavy-lifting of making a viable study.
-- [`study_generation`](study_generation/): High level of each type of file and each pipeline it can process.
-- [`data_type`](data_type/): Actual processing of each piece of data
-- [`constants`](constants/) & [`support`](support/): Other functions and data used by above 2
-- [`tools`](tools/): Folder holding the `main` programs for all 4 of Janus' faces
